Feat: 팝업 데이터 연결 완료
